+++ apps/io-session-manager-internal/src/__mocks__/platform-internal-client.mock.ts
@@ -1,0 +1,11 @@
+import { vi } from "vitest";
+import * as E from "fp-ts/lib/Either";
+
+export const mockDeleteSession = vi.fn().mockResolvedValue(
+  E.right({
+    status: 204,
+  }),
+);
+export const mockPlatformInternalClient = {
+  deleteSession: mockDeleteSession,
+};