+++ apps/io-session-manager-internal/CHANGELOG.md
@@ -1,5 +1,17 @@
 # io-session-manager-internal
 
+## 0.17.1
+
+### Patch Changes
+
+- 0a529ba: Fix: invalidateUserSession idempotency
+
+## 0.17.0
+
+### Minor Changes
+
+- bfaaf17: Implemented session token cache invalidation
+
 ## 0.16.0
 
 ### Minor Changes